refactor : replace useState with custom hooks 'useValidation' in Input & update type
